refactor: move Hotwire importmaps and assets to extension system
- Move Hotwire-related importmaps from config defaults to a dedicated extension
- Move Hotwire assets to extension directory and update asset path logic
- Refactor asset lookup to use site instance methods instead of Hotpages class methods
- Update tests and expected outputs for new importmap and asset ordering
