+++ lib/hotpages.rb
@@ -23,7 +23,7 @@ module Hotpages
     end
 
     def extensions = @extensions ||= [
-      Extensions::I18n
+      Extensions::I18n, Extensions::Hotwire
     ]
     def remove_extension(extension) = extensions.delete(extension)
     def extension_helpers = @extension_helpers ||= []
@@ -43,18 +43,7 @@ module Hotpages
     end
 
     # TODO: move assets related methods to Site
-    def assets_path = File.join(__dir__, "hotpages/assets")
-    def assets_paths = [ assets_path, site.assets_path ].compact
-    def assets(filter_ext = nil)
-      Enumerator.new do |yielder|
-        assets_paths.each do |path|
-          Dir.glob(File.join(path, "**", "*#{filter_ext}")).select do |file|
-            next unless File.file?(file)
-            yielder << [ path, file ]
-          end
-        end
-      end
-    end
+    # def assets_path = File.join(__dir__, "hotpages/assets")
 
     def dev_server
       raise "Site is not set. Please call Hotpages.setup_site first." unless site
