+++ lib/hotpages/extensions/hotwire/stimulus-loading.js
